Improve names and documentation.
