+++ src/lib.rs
@@ -13,7 +13,7 @@
 //  
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#![feature(io)]
+#![feature(io, iter_arith)]
 
 extern crate base64;
 extern crate mime;
