+++ src/terminal/screen/mod.rs
@@ -1,19 +1,18 @@
 use std::ops::{Deref, DerefMut};
 
-use datatypes::{Region, SaveGrid, SplitKind, ResizeRule};
-use terminal::char_grid::CharGrid;
+use datatypes::{Region, SaveGrid, SplitKind, ResizeRule, CoordsIter};
+use terminal::char_grid::{CharGrid, CharCell};
 
-mod stack;
 mod panel;
+mod section;
+mod stack;
 
-use self::panel::Panel;
-
-pub use self::panel::{Cells, Panels};
-pub use self::stack::Stack;
+use self::section::ScreenSection;
+use self::panel::Panel::*;
 
 pub struct Screen {
     active: u64,
-    screen: Panel,
+    screen: ScreenSection,
 }
 
 impl Screen {
@@ -21,41 +20,41 @@ impl Screen {
     pub fn new(width: u32, height: u32) -> Screen {
         Screen {
             active: 0,
-            screen: Panel::new(0, Region::new(0, 0, width, height)),
+            screen: ScreenSection::new(0, Region::new(0, 0, width, height)),
         }
     }
 
     pub fn resize(&mut self, width: Option<u32>, height: Option<u32>, rule: ResizeRule) {
         let new_a = match (width, height) {
             (Some(w), Some(h))  => Region::new(0, 0, w, h),
-            (Some(w), None)     => Region::new(0, 0, w, self.screen.area.bottom),
-            (None,    Some(h))  => Region::new(0, 0, self.screen.area.right, h),
+            (Some(w), None)     => Region::new(0, 0, w, self.screen.area().bottom),
+            (None,    Some(h))  => Region::new(0, 0, self.screen.area().right, h),
             (None,    None)     => return
         };
         self.screen.resize(new_a, rule);
     }
 
     pub fn switch(&mut self, tag: u64) {
-        if self.find(Some(tag)).map_or(false, Panel::is_grid) {
+        if self.find(Some(tag)).map_or(false, ScreenSection::is_grid) {
             self.active = tag;
         }
     }
 
     pub fn split(&mut self, save: SaveGrid, kind: SplitKind, rule: ResizeRule,
                  split_tag: Option<u64>, l_tag: u64, r_tag: u64) {
-        self.find_mut(split_tag).map(|panel| panel.split(save, kind, rule, l_tag, r_tag));
+        self.find_mut(split_tag).map(|section| section.split(save, kind, rule, l_tag, r_tag));
     }
 
     pub fn unsplit(&mut self, save: SaveGrid, unsplit_tag: Option<u64>) {
-        self.find_mut(unsplit_tag).map(|panel| panel.unsplit(save));
+        self.find_mut(unsplit_tag).map(|section| section.unsplit(save));
     }
 
     pub fn push(&mut self, tag: Option<u64>) {
-        self.find_mut(tag).map(Panel::push);
+        self.find_mut(tag).map(ScreenSection::push);
     }
 
     pub fn pop(&mut self, tag: Option<u64>) {
-        self.find_mut(tag).map(Panel::pop);
+        self.find_mut(tag).map(ScreenSection::pop);
     }
 
     pub fn cells(&self) -> Cells {
@@ -66,11 +65,11 @@ impl Screen {
         self.screen.panels()
     }
 
-    fn find(&self, tag: Option<u64>) -> Option<&Panel> {
+    fn find(&self, tag: Option<u64>) -> Option<&ScreenSection> {
         self.screen.find(tag.unwrap_or(self.active))
     }
 
-    fn find_mut(&mut self, tag: Option<u64>) -> Option<&mut Panel> {
+    fn find_mut(&mut self, tag: Option<u64>) -> Option<&mut ScreenSection> {
         self.screen.find_mut(tag.unwrap_or(self.active))
     }
 
@@ -89,3 +88,74 @@ impl DerefMut for Screen {
     }
 }
 
+/// An iterator over all of the cells in a section of the screen - either one panel or the entire
+/// screen.
+pub struct Cells<'a> {
+    iter: CoordsIter,
+    screen: &'a ScreenSection,
+}
+
+impl<'a> Cells<'a> {
+    /// The section of the screen that this iterator iterates over.
+    pub fn area(&self) -> Region {
+        self.iter.region()
+    }
+}
+
+impl<'a> Iterator for Cells<'a> {
+    type Item = &'a CharCell;
+
+    fn next(&mut self) -> Option<&'a CharCell> {
+        self.iter.next().map(|coords| &self.screen[coords])
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.iter.size_hint()
+    }
+
+}
+
+impl<'a> DoubleEndedIterator for Cells<'a> {
+    fn next_back(&mut self) -> Option<&'a CharCell> {
+        self.iter.next_back().map(|coords| &self.screen[coords])
+    }
+}
+
+impl<'a> ExactSizeIterator for Cells<'a> { }
+
+/// An iterator over all of the visible panels in the terminal's screen.
+pub struct Panels<'a> {
+    stack: Vec<&'a ScreenSection>,
+}
+
+impl<'a> Iterator for Panels<'a> {
+    type Item = Cells<'a>;
+
+    fn next(&mut self) -> Option<Cells<'a>> {
+        fn cells<'a>(section: &'a ScreenSection, stack: &mut Vec<&'a ScreenSection>) -> Cells<'a> {
+            match *section.top() {
+                Grid(_) => Cells {
+                    iter: CoordsIter::from_region(section.area()),
+                    screen: section
+                },
+                Split { ref left, ref right, .. } => {
+                    stack.push(right);
+                    cells(left, stack)
+                }
+                _ => unreachable!()
+            }
+        }
+        self.stack.pop().map(|section| cells(section, &mut self.stack))
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.len(), Some(self.len()))
+    }
+
+}
+
+impl<'a> ExactSizeIterator for Panels<'a> {
+    fn len(&self) -> usize {
+        self.stack.iter().cloned().map(ScreenSection::count_grids).sum()
+    }
+}
